+++ eras/babbage/impl/src/Cardano/Ledger/Babbage/Translation.hs
@@ -158,6 +158,7 @@ instance Crypto c => TranslateEra (BabbageEra c) ShelleyGovState where
         , sgsFutureProposals = translateEra' ctxt $ sgsFutureProposals ps
         , sgsCurPParams = translateEra' ctxt $ sgsCurPParams ps
         , sgsPrevPParams = translateEra' ctxt $ sgsPrevPParams ps
+        , sgsFuturePParams = translateEra' ctxt <$> sgsFuturePParams ps
         }
 
 instance Crypto c => TranslateEra (BabbageEra c) ProposedPPUpdates where