+++ eras/shelley/impl/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.11.0.0
 
-*
+* Introduce `futurePParamsGovStateL`, `futurePParamsShelleyGovStateL` and `sgsFuturePParams`
 
 ### `testlib`